Add in black level correction.
The lens shading compensation is done after black level correction
in the ISP pipe, therefore the raw data being processed to generate
the grid needs to have black level correction applied.

Add this in, taking an optional command line parameter to set the
black level.
For reference, ov5647 is 16 (the default), and imx219 is 66.

Credit to rwb27 for realising this
https://github.com/6by9/lens_shading/issues/2